+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -7,13 +7,12 @@ import org.jgrapht.graph.WeightedMultigraph;
 import org.springframework.stereotype.Component;
 
 import nextstep.subway.line.domain.Section;
-import nextstep.subway.line.domain.TotalLines;
 import nextstep.subway.station.domain.Station;
 
 @Component
 public class PathFinder {
 
-	public ShortestPath findShortestPath(TotalLines lines, Station sourceStation, Station targetStation) {
+	public ShortestPath findShortestPath(SubwayMap lines, Station sourceStation, Station targetStation) {
 		DijkstraShortestPath dijkstraShortestPath = this.findDijkstraShortestPath(lines);
 		GraphPath path = dijkstraShortestPath.getPath(sourceStation, targetStation);
 
@@ -24,7 +23,7 @@ public class PathFinder {
 		return new ShortestPath(path.getVertexList(), Math.round(path.getWeight()));
 	}
 
-	private DijkstraShortestPath findDijkstraShortestPath(TotalLines lines) {
+	private DijkstraShortestPath findDijkstraShortestPath(SubwayMap lines) {
 		WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph(DefaultWeightedEdge.class);
 
 		for(Section section : lines.allSections()) {
